requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sse-starlette==2.0.0
prometheus-fastapi-instrumentator==6.1.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.0
cachetools==5.3.2
loguru==0.7.2